+++ democracylab/settings.py
@@ -452,6 +452,7 @@ QIQO_API_SECRET = os.environ.get('QIQO_API_SECRET', 'SECRET')
 QIQO_CIRCLE_UUID = os.environ.get('QIQO_CIRCLE_UUID', 'nmitq')
 QIQO_SIGNUP_TIMEOUT_SECONDS = int(os.environ.get('QIQO_SIGNUP_TIMEOUT_SECONDS', 5))
 
+SALESFORCE_CONNECTED = True
 SALESFORCE_API_VERSION = os.environ.get('SALESFORCE_API_VERSION')
 SALESFORCE_ENDPOINT = os.environ.get('SALESFORCE_ENDPOINT')
 SALESFORCE_LOGIN_URL = os.environ.get('SALESFORCE_LOGIN_URL')
